chore: move BIRBuilder to birgen
